# sync opt