--- a/rust/automerge/src/sync/message_builder.rs
+++ b/rust/automerge/src/sync/message_builder.rs
@@ -71,8 +71,4 @@ impl MessageBuilder {
             version: self.version,
         }
     }
-
-    pub(super) fn has_changes_to_send(&self) -> bool {
-        !self.changes.is_empty()
-    }
 }