Two big optimizations for sync

1. lots of the sync code was still requesting full changes when all it needed was hashes (2.2 -> 3.0) issue.  This has been fixed
2. found a bug where the send_whole doc was not triggering most of the time leading to much slower init sync

my test sync for 
 full -> full  went form 1600ms -> 17ms 
 full -> empty went from 3400ms -> 1600ms